[Phi] Migrate matrix_solve to phi (#44298)
* [Phi] Migrate matrix_solve to phi

* replace mutable_data with Alloc
