+++ paddle/phi/kernels/funcs/CMakeLists.txt
@@ -14,3 +14,4 @@ math_library(matrix_inverse DEPS dense_tensor eigen3 blas)
 math_library(pooling DEPS dense_tensor)
 math_library(segment_pooling)
 math_library(sequence2batch)
+math_library(matrix_solve DEPS dense_tensor eigen3 blas math_function)
